Reorient on how to map the entire via chain - again.
Hibernate's collection mapping and I aren't getting along at the moment
so for now we'll do this manually.  In addition, the code, as currently
written, will create duplicate via chains.  I need to add in a reuse
check.

And tests.  Lots and lots and lots of tests.

Plus clean up all of the cruft that has been accumulating.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@929 bc24b658-cf39-4249-b827-e3526eaaeb42