Removed all references to Sarissa, the cross-browser-compatibility Javascript package. It turned out to be overkill for our needs.
We're no longer sending XML in response to an XMLHTTPRequest. Instead, we're justing sending text, which we later insert into the page as HTML. This approach is more compatible with more browsers, and simplfies our JavaScript code considerably.

Personview.jsp no longer counts on receiving a non-NULL currentSubsystem value. This flexibility is required for persons who do not yet have any Assignments in any Subsystems.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@1012 bc24b658-cf39-4249-b827-e3526eaaeb42