Assignment.getLimitValues now more reliably sets the Signet member in each of the returned Limits.
Removed some obsolete references from PrivilegedSubjectImpl.java.

Extended PrivilegedSubject.canEdit() and PrivilegedSubject.editRefusalExplanation() to take Limit-values into account when deciding whether or not a PrivilegedSubject can edit a particular Assignment.

Added a new JUnit test: PrivilegedSubjectTest.testCanEdit().

Checkboxes for Assignment-revocation are now greyed out in accordance with the new canEdit() logic, and the reason for a greyed-out checkbox is available as a tool-tip associated with the checkbox.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@1057 bc24b658-cf39-4249-b827-e3526eaaeb42