Lowered the default Signet logging level from "info" to "error". This quiets down the utility programs considerably.
Changed the build process for SubjectFileLoader to more closely match the deployments of TreeXMLLoader and SubsystemXMLLoader.

Gave SubjectFileLoader a "run.sh" file to match those associated with TreeXMLLoader and SubsystemXMLLoader.

Changed the line-terminations of the existing "run.sh" files to Unix-style (newline only) from Windows-style (carriage-return and newline). The Windows-style line-terminators were preventing those scripts from executing on Unix.

Made some more changes to the README file to make it more readable.

Changed the order of operations in the README file to reflect the fact that Trees must be loaded before Subsystems.

Fixed a typo in the README file's sample subject-loader invocation.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@1210 bc24b658-cf39-4249-b827-e3526eaaeb42