Fixed bug 3086:
RevokeAction.java and RevokeAndGrantAction.java were both lacking a call to Assignment.save(). I added those missing invocations.

Fleshed out the implementation of Assignment attributes "revokerID" and "revokerTypeID". It's important to note that while an Assignment's grantor and grantee can never be null, it's OK for an Assignment to have a null revoker, if the Assignment has never been revoked.

Removed a redundant Signet.save() call from AssignmentImpl.save().

Renamed Signet.getSource() to Signet.getSources(), to better reflect the fact that it returns a Collection of Sources.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@1406 bc24b658-cf39-4249-b827-e3526eaaeb42