PersonViewAction is now just a subject-neutral version of StartAction, which always resets the current subject to the logged-in user.
main.jsp now displays the "start granting privilege" button when the displayed subject is not the logged-in user.

main.jsp received some more JavaScript code to handle the "start granting privilege" button, which was formerly handled only by personview.jsp.

The "change the view to show" button now works more correctly, preserving the notion of "current subject" instead of re-setting it to the logged-in user.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@1575 bc24b658-cf39-4249-b827-e3526eaaeb42