Grantable.evaluate() now returns a boolean, indicating whether or not the call changed the Status of the Grantable object.
New method Grantable.evaluate(date) allows the caller to specify an evaluation-date. This is useful for testing, and for causing all Grantable objects to be evaluated as of the exact same time.

ProxyImpl.findDuplicates() now examines in-memory values instead of performing a SQL query, because the Hibernate-SQL query was not handling NULL Subsystem-IDs properly.

JUnit tests were updated to accommodate the change in Grantable.evaluate().


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@2215 bc24b658-cf39-4249-b827-e3526eaaeb42