Completely rewrote this class, retaining functionality, which is good.
Use a Map mapping lowercased values to actual values instead of using BasicAttribute. BasicAttribute has no case-indistinct operations, making a linear search necessary to find and/or remove attributes. This doesn't scale well if groups have tens or hundreds of thousands of members.

Using a map (which becomes an identity map if the attribute is case distinct) allows for hash-time searches, which should be log(n).

Implemented the maps with an interior class to keep the main code clean and clear.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@4303 bc24b658-cf39-4249-b827-e3526eaaeb42