GRP-1588: Ldap Groups - "Entry already exists" errors when groups found by DN/EntryDn/DistinguishedName
Added automatic recovery to PSPNG when bulk searching for groups fails. This can particularly happen
when singleGroupSearchFilter is configured to look for groups by DN because DNs can be esacped
and the necessary in-memory group search does not work with escaped DNs.

To avoid the warnings (see below) and duplicate ldap searches, this commit adds an enableBulkGroupSearching config
property that disables the bulk searching and always does slower-but-simpler, single-group searches.

WARN  Bulk fetch failed (returned unmatchable group data). This can be caused by searching for a DN with escaping or by singleGroupSearchFilter ... that are not included in groupSearchAttributes
WARN  Slower fetching will be attempted